+++ examples/Cargo.toml
@@ -18,6 +18,7 @@ futures = "0.3.21"
 indicatif = "0.17.0"
 iroh-car = { version = "0.1.3", path = "../iroh-car" }
 iroh-resolver = { version = "0.1.3", path = "../iroh-resolver" }
+iroh-content = { version = "0.1.3", path = "../iroh-content" }
 iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client" }
 iroh-util = { version = "0.1.3", path = "../iroh-util" }
 par-stream = { version = "0.10.2", default-features = false, features = ["runtime-tokio"] }
