+++ iroh-gateway/src/bad_bits.rs
@@ -111,7 +111,7 @@ mod tests {
     use super::*;
     use hex_literal::hex;
     use http::StatusCode;
-    use iroh_resolver::content_loader::{FullLoader, FullLoaderConfig, GatewayUrl};
+    use iroh_content_loader::content_loader::{FullLoader, FullLoaderConfig, GatewayUrl};
     use iroh_resolver::dns_resolver::Config as DnsResolverConfig;
     use iroh_rpc_client::{Client as RpcClient, Config as RpcClientConfig};
