+++ iroh-gateway/src/core.rs
@@ -1,5 +1,5 @@
 use axum::Router;
-use iroh_resolver::content_loader::ContentLoader;
+use iroh_content::content_loader::ContentLoader;
 use iroh_rpc_types::gateway::GatewayAddr;
 
 use iroh_resolver::dns_resolver::Config as DnsResolverConfig;
@@ -122,13 +122,13 @@ mod tests {
     use super::*;
     use cid::Cid;
     use futures::{StreamExt, TryStreamExt};
-    use iroh_resolver::content_loader::{FullLoader, FullLoaderConfig};
-    use iroh_resolver::unixfs::UnixfsNode;
-    use iroh_resolver::unixfs_builder::{DirectoryBuilder, FileBuilder};
+    use iroh_content::content_loader::{FullLoader, FullLoaderConfig};
     use iroh_rpc_client::Client as RpcClient;
     use iroh_rpc_client::Config as RpcClientConfig;
     use iroh_rpc_types::store::StoreAddr;
     use iroh_rpc_types::Addr;
+    use iroh_unixfs::builder::{DirectoryBuilder, FileBuilder};
+    use iroh_unixfs::unixfs::UnixfsNode;
     use std::io;
     use tokio_util::io::StreamReader;
