+++ iroh-gateway/src/main.rs
@@ -2,6 +2,7 @@ use std::sync::Arc;
 
 use anyhow::{anyhow, Context, Result};
 use clap::Parser;
+use iroh_content::content_loader::{FullLoader, FullLoaderConfig};
 use iroh_gateway::{
     bad_bits::{self, BadBits},
     cli::Args,
@@ -9,7 +10,6 @@ use iroh_gateway::{
     core::Core,
     metrics,
 };
-use iroh_resolver::content_loader::{FullLoader, FullLoaderConfig};
 use iroh_rpc_client::Client as RpcClient;
 use iroh_util::lock::ProgramLock;
 use iroh_util::{iroh_config_path, make_config};
