+++ iroh-one/Cargo.toml
@@ -24,6 +24,7 @@ iroh-gateway = { version = "0.1.3", path = "../iroh-gateway" }
 iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features = false }
 iroh-p2p = { version = "0.1.3", path = "../iroh-p2p", default-features = false }
 iroh-resolver = { version = "0.1.3", path = "../iroh-resolver" }
+iroh-content = { version = "0.1.3", path = "../iroh-content" }
 iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
 iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
 iroh-store = { version = "0.1.3", path = "../iroh-store", default-features = false }
