+++ iroh-one/src/main.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 #[allow(unused_imports)]
 use anyhow::{anyhow, Result};
 use clap::Parser;
+use iroh_content::content_loader::{FullLoader, FullLoaderConfig};
 use iroh_gateway::{bad_bits::BadBits, core::Core, metrics};
 #[cfg(all(feature = "http-uds-gateway", unix))]
 use iroh_one::uds;
@@ -10,7 +11,6 @@ use iroh_one::{
     cli::Args,
     config::{Config, CONFIG_FILE_NAME, ENV_PREFIX},
 };
-use iroh_resolver::content_loader::{FullLoader, FullLoaderConfig};
 use iroh_rpc_client::Client as RpcClient;
 use iroh_rpc_types::Addr;
 use iroh_util::lock::ProgramLock;
