+++ iroh-resolver/Cargo.toml
@@ -28,6 +28,7 @@ iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features =
 iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
 iroh-util = { version = "0.1.3", path = "../iroh-util", default-features = false }
 iroh-content = { path = "../iroh-content" }
+iroh-unixfs = { version = "0.1.3", path = "../iroh-unixfs" }
 libipld = "0.15.0"
 libp2p = { version = "0.50", default-features = false, features = ["serde"] }
 multihash = "0.17"
