+++ iroh-resolver/tests/unixfs.rs
@@ -2,11 +2,9 @@ use std::{io::Read, path::PathBuf, time::Instant};
 
 use anyhow::Result;
 use bytes::Bytes;
-use iroh_resolver::{
-    chunker::{self, Chunker},
-    resolver::{OutMetrics, Path, ResponseClip},
-    unixfs_builder::{read_to_vec, stream_to_resolver, FileBuilder},
-};
+use iroh_content::chunker::{self, Chunker};
+use iroh_content::content::{OutMetrics, Path, ResponseClip};
+use iroh_unixfs::builder::{read_to_vec, stream_to_resolver, FileBuilder};
 
 async fn read_fixture(path: impl AsRef<std::path::Path>) -> Result<Vec<u8>> {
     let path = path.as_ref().to_owned();
