+++ iroh-share/src/p2p_node.rs
@@ -3,12 +3,13 @@ use std::{collections::HashSet, path::Path, sync::Arc};
 use anyhow::{ensure, Result};
 use async_trait::async_trait;
 use cid::Cid;
-use iroh_p2p::{config, Keychain, MemoryStorage, NetworkEvent, Node};
-use iroh_resolver::{
-    content_loader::ContentLoader,
-    parse_links,
-    resolver::{ContextId, LoadedCid, LoaderContext, Resolver, Source, IROH_STORE},
+use iroh_content::{
+    content::{LoadedCid, Source},
+    content_loader::{ContentLoader, ContextId, LoaderContext, IROH_STORE},
+    util::parse_links,
 };
+use iroh_p2p::{config, Keychain, MemoryStorage, NetworkEvent, Node};
+use iroh_resolver::resolver::Resolver;
 use iroh_rpc_client::Client;
 use iroh_rpc_types::Addr;
 use libp2p::{Multiaddr, PeerId};
