+++ iroh-unixfs/src/lib.rs
@@ -1,4 +1,4 @@
-mod balanaced_tree;
+pub mod balanced_tree;
+pub mod builder;
 pub mod hamt;
 pub mod unixfs;
-pub mod unixfs_builder;
